Ejercicio1/requirements.txt
pyyaml
pytest
